requirements.txt
fastapi==0.98.0
uvicorn==0.22.0
python-decouple==3.8
SQLAlchemy==2.0.17

